+++ PokerGameApp/PokerGameApp/Player.swift
@@ -5,7 +5,7 @@ class Player: CustomStringConvertible{
     var currentName: String{
         return self.name
     }
-    fileprivate (set) var cards: Cards
+    private (set) var cards: Cards
     var description: String{
         return "\(name) \(cards)"
     }
@@ -17,22 +17,6 @@ class Player: CustomStringConvertible{
     
     func addCard(_ card: Card){
         self.cards.addCard(card)
-    }
-}
-
-class Dealer: Player{
-    
-    init() {
-        super.init(name: "딜러")
-    }
-    
-    func takeOutAllNecessaryCards(deck: CardDeck, count: Int)->CardDeck{
-        var deckCopy = deck
-        for _ in 0..<count{
-            guard let card = deckCopy.removeOne() else { continue }
-            super.addCard(card)
-        }
-        return deckCopy
     }
     
     func takeOutCard()-> Card?{
